+++ js/util.js
@@ -52,9 +52,6 @@ const isEscapeKey = (evt) => evt.key === 'Escape';
 // Функция для проверки нажатия на кнопку Enter
 const isEnterKey = (evt) => evt.key === 'Enter';
 
-// Функция для проверки нажатия на кнопку "закрыть" модального окна
-const isCloseButton = (evt) => evt.closest('.big-picture__cancel');
-
 // Функция для нахождения номера выбранного элемента коллекции
 const findElementNumber = (element, collection) => {
   for (let currentElement = 0; currentElement < collection.length; currentElement++) {
@@ -64,4 +61,4 @@ const findElementNumber = (element, collection) => {
   }
 };
 
-export {getRandomNumber, limitCommentaryLength, getRandomArrayElement, createObjectsArray, createUniqueRandomNumber, isEscapeKey, isEnterKey, isCloseButton, findElementNumber};
+export {getRandomNumber, limitCommentaryLength, getRandomArrayElement, createObjectsArray, createUniqueRandomNumber, isEscapeKey, isEnterKey, findElementNumber};
